refactor: 초기에 두 장의 카드를 받는 기능을 Role으로 이동
