+++ src/main/kotlin/blackjack/model/participant/Role.kt
@@ -1,6 +1,7 @@
 package blackjack.model.participant
 
 import blackjack.model.card.Card
+import blackjack.model.card.CardProvider
 
 abstract class Role {
     private val handCards = HandCards()
@@ -9,6 +10,10 @@ abstract class Role {
 
     fun receiveCard(cards: List<Card>) {
         handCards.add(cards)
+    }
+
+    fun initCard(cardProvider: CardProvider) {
+        receiveCard(Card.provideCards(cardProvider, INIT_RECEIVE_CARD_COUNT))
     }
 
     fun isBurst() = handCards.getCardSum(BLACKJACK_VALUE) > BLACKJACK_VALUE
@@ -21,5 +26,6 @@ abstract class Role {
 
     companion object {
         private const val BLACKJACK_VALUE = 21
+        private const val INIT_RECEIVE_CARD_COUNT = 2
     }
 }
